devel/xsd: Update to 4.2.0
Add two new support ports: devel/libxsd and devel/libxsd-frontend.

The runtime library was moved out of devel/xsd to devel/libxsd which is
why the headers are gone. The compiler frontend, which following a
previous comment in the devel/xsd Makefile suggesting that it should be a
separate port, but more out of necessity for compatibility with the
build2 build system, is now devel/libxsd-frontend.

The examples are now a separate distribution and I didn't really see any
urgent need to reincorporate them, as they would require yet another port
that the majority of our userbase would probably not be interested in as
xsd is ultimately just a build dependency for textproc/libkolabxml.

Update devel/libcutl to 1.11.0 due to header relocation.

Update textproc/libkolabxml to HEAD for C++11 compatibility.

Differential Revision:	https://reviews.freebsd.org/D43080