textproc/jade: remove non-functional deoptimization
Long ago, this code was added to disable -O2 optimizations and
presumably remove -mcpu flags to avoid miscompilation.  Somewhere along
the line it stopped working and code is being compiled with -O2 by
default and proably with -mcpu flags if set.  The only place this seems
to be used is when checking if the compiler works early in configure.

Discovered on CheriBSD where removing -m* removed created an invalid
command line with two -Xclang flags in a row.

Reviewed by:	cy
Sponsored by:	DARPA
Differential Revision:	https://reviews.freebsd.org/D43072