x11/xidle: fix double triggering on screen unlock
A problem reported that xidle triggers twice: the first one when it
reaches idle timeout and locks the screen (which is expected), and the
second time immediately when the screen is unlocked (which is
undesired).
The problem is that, apparently, the XScreenServer extension, which
xidle uses, sends ScreenSaverOn on locking and ScreenSaverOff on
unlocking. Currently, xidle doesn't filter events by locking state, so
it locks the screen again when the ScreenSaverOff event comes in.

The fix is to run a screensaver only on the ScreenSaverOn event.

PR:	275761