java/wildfly: New port
Replacement for JBoss Application Server

WildFly is a flexible, lightweight, managed application runtime that
helps you build amazing applications.  WildFly - new name for JBoss
Application Server

- Fast Startup
- Small Footprint
- Modular Design
- Unified Configuration and Management
And of course Java EE!

WWW: https://wildfly.org/

As the current MAINTAINER is no longer adding any support and all
wildfly versions are about to EXPIRE add this port. Please be informed
that this is still an older version of the port. There was a breaking
change in wildfly which I cannot recollect so I will need to test and
update. But in case someone quickly recollects the breaking version
number give me a heads up. Otherwise I will test and update soonish.