graphics/nvidia-drm-515-kmod: Fix build on CURRENT with iosys-map.h
Commit f8ab2f5bae1d added the linux/iosys-map.h header to base in order
to support 5.18+ versions of drm-kmod. This messes with nvidia-drm's
conftest logic as on recent enough CURRENT versions it will see iosys-map
and assume it can use it, but the drm-kmod version in use is 5.15,
which wants to use the older dma_buf_map.

This adds a patch on the port side to override nvidia-drm's detector
and force the use of dma_buf_map on the 5.15 port.

Differential Revision:	https://reviews.freebsd.org/D43234