KDE: Update KDE Frameworks 6 to 5.247.o (beta 2)
Plasma 6, Frameworks and Gear draw closer

Every few years we port the key components of our software to a new
version of Qt, taking the opportunity to remove cruft and leverage the
updated features the most recent version of Qt has to offer us.

We are now just over two months away from KDE's megarelease. At the end
of February 2024 we will publish Plasma 6, Frameworks 6 and a whole new
set of applications in a special edition of KDE Gear all in one go.

If you have been following the updates here and here, you will know we
are deep in the testing phase; and KDE is making available today the
second Beta version of all the software we will include in the
megarelease.

As with versions Alpha and the first Beta, this is a preview intended
for developers and testers. The software in this second beta release is
reaching stability fast, but it is still not 100% safe to use in a
production environment. We still recommend you continue using stable
versions of Plasma, Frameworks and apps for your everyday work. But if
you do use this, watch out for bugs and report them promptly, so we can
solve them.

Read on to find out more about KDE's 6th Megarelease, what it covers,
and how you can help make the new versions of Plasma, KDE's apps and
Frameworks a success now.

Announcement: https://kde.org/announcements/megarelease/6/beta2/