biology/ucsc-userapps: Replace dependencies
Currently their is a hardcoded dependency on mysql57-client which is
about to be removed. Ports depending on mysql should always use
USES=mysql rather than hardcoded *_DEPENDS. As I do not have a licensed
version of this application for now switch to USES=mysql. This should
not affect the cluster build at all as this is a RESTRICTED ports.

Approved by:	portmgr (blanket)