KDE: Update KDE Plasma Desktop 6 to 5.91.0 (beta 2)
Plasma is KDE's flagship desktop environment. Plasma is like Windows or
macOS, but is renowned for being flexible, powerful, lightweight and
configurable. It can be used at home, at work, for schools and research.

Plasma 6 is the upcoming version of Plasma that integrates the latest
version of Qt, Qt 6, the framework upon which Plasma is built.

Plasma 6 incorporates new technologies from Qt and other constantly
evolving tools, providing new features, better support for the latest
hardware, and supports for the hardware and software technologies to
come.

Announcement:	https://kde.org/announcements/megarelease/6/beta2/