devel/kommit: Update to HEAD (1.3.1-20)
Version 1.3.1 was tagged, but distfiles ever since 1.0.2 have not been
created as of writing for unknown reasons, so fetch from the upstream
git repo instead. There exists a rather nasty bug that disrupts the
space-time continuum in the v1.3.1 tag, however, where displayed commit
dates are off by a factor of 10^-3, displaying milliseconds instead of
seconds elapsed since the UNIX epoch. This has since been fixed in [1],
but a new tag that includes this commit has yet to be created, so I'm
just going with the latest commit which seems to build and run fine.
Most of the other commits after the v1.3.1 tag are bot-created to update
documentation and the human-created ones appear to be entirely
reasonable.

The main purpose of this update is to try to mitigate crashes with large
and long-historied codebases like the FreeBSD src tree as reported in
[2]. Upstream has migrated to using devel/libgit2 since the previous
version (1.0.2) which makes kommit load very large trees noticably faster
(anecdotally, 8-9 seconds faster for the entire FreeBSD src tree on my 4
year old laptop which is still rather slow at around 21 seconds,
at least by modern expectations of instantaneous results, but a win is
still a win).

Plus it seems to be more stable in terms of closing cleanly on exit.
For example, 1.0.2, likely unbeknownst to the user, would continue to
consume roughly 4GB of RAM and happily burden 50% of a CPU core. It may
have actually closed eventually, but after 5 minutes I got borded and
terminated the process manually. I'm no longer seeing this behavior in
1.3.1-20, however. The application now exits almost immediately once the
tree has been loaded and either window close or choosing Quit from the
file menu (Ctrl+Q) has been selected.

[1] https://invent.kde.org/sdk/kommit/-/commit/6d0a0bb043de981c3b4d7fe54debe49ead541dbe

PR:		271683 [2]
Reported by:	grahamperrin [2]