net-im/psi: Convert WEBKIT option to WEBENGINE
www/qt5-webkit was removed in [1], leaving the WEBKIT option broken.
This application supports using www/qt5-webengine as an alternative
chatlog engine, so rename the WEBKIT option to WEBENGINE to avoid
confusion and add the required dependencies. Enabling QtWebEngine still
requires having the ENABLE_WEBKIT CMake argument set, so this was not an
oversight [2], as strange as it may look. In fact, psi defaults to use
QtWebEngine if it was found when the ENABLE_WEBKIT argument is set. [3]

[1] https://cgit.freebsd.org/ports/commit/?id=cb3615b528c288d90aff6aa0e7a9778f80c9e85d
[2] https://github.com/psi-im/psi/blob/1.5/CMakeLists.txt#L44
[3] https://github.com/psi-im/psi/blob/1.5/src/CMakeLists.txt#L17

Approved by:	portmgr (blanket)
MFH:		2024Q1