graphics/digikam: Remove WEBENGINE option.
After removal of www/qt5-webkit, www/qt5-webengine is the only
alternative.

While here, remove the dependency on graphics/liblqr-1 which has been
non-optionally bundled starting with 8.2.0 release. The Akonadi patches
need some work, but it builds, so I'll save that for the next release
along with a few other cleanups. Not depending on a non-existent port
is more pressing at this time.

MFH:		2024Q1