x11/libxfce4util: Workaround gettext library detection issue
Due to an issue in the configure script gettext functions are not
detected properly t configure time.

This is causing display issues with non-UTF8 locales.

This commit implements a workaround by passing some extra env to
configure to successfully detect the library.

A proper fix is proposed upstream at [1].

[1] https://gitlab.xfce.org/xfce/libxfce4util/-/merge_requests/41

PR:		276211