KDE: Update KDE Frameworks 6 to 5.248
KDE's 6th Megarelease - Release Candidate 1

Wednesday, 10 January 2024

Plasma 6, Frameworks and Gear draw closer

Every few years we port the key components of our software to a new
version of Qt, taking the opportunity to remove cruft and leverage the
updated features the most recent version of Qt has to offer us.

KDE's megarelease is less than 50 days away. At the end of February 2024
we will publish Plasma 6, Frameworks 6, and a whole new set of
applications in a special edition of KDE Gear all in one go.

If you have been following the updates here, here and here, you will
know we are making our way through the testing phase and gradually
reaching stability. KDE is making available today the first Release
Candidate version of all the software we will include in the
megarelease.

As with the Alpha and Beta versions, this is a preview intended for
developers and testers. The software provided is nearing stability, but
is still not 100% safe to use in a production environment. We still
recommend you continue using stable versions of Plasma, Frameworks and
apps for your everyday work. But if you do use this, watch out for bugs
and report them promptly, so we can solve them.

Read on to find out more about KDE's 6th Megarelease, what it covers,
and how you can help make the new versions of Plasma, KDE's apps and
Frameworks a success now.

To get a full Qt6 based desktop, the overlay below can be used (once it
is updated to RC1)

Announcement:	https://kde.org/announcements/megarelease/6/rc1/
Overlay:	https://github.com/tcberner/kde6-overlay/