databases/sqldeveloper: make depends from java/openjfx14 mandatory
The SQLDeveloper distribution has built-in OpenJFX for other OSes,
and it can work with it on FreeBSD (by forcing the path to one of them),
but not all functions work.
We have very little reason to keep this option, then we have a fully
functional port java/openjfx14.

Approved by:	arrowd (mentor)