devel/appstream{,-qt,-compose}: Combine into a single port with SUBPACKAGES.
The Qt part was a bit complex as it had two FLAVORS. To work this out the
slave devel/appstream-qt port wasn't removed but instead switched to build
the master port without SUBPACKAGES and against Qt5 only.

The Qt subpackage of the master port is fixed to Qt6 instead. This will make it
possible to just remove Qt5 slave port when we get rid of Qt5 altogether.

Reviewed by: mat

Differential Revision: https://reviews.freebsd.org/D43445