net/dpdk-20.11: Fix build warnings about symbols in version map file
The latest versions of ld are throwing warnings/errors about entries in
the linker script that don't exist in the actual build. For DPDK, there
are functions in the linker script (version.map) which only exist on Linux
(and sometimes only on Windows) but not on BSD. To avoid warnings/errors,
we can restore old behaviour by adding the "-Wl,--undefined-version"
flag to the build.

PR:		276277