net/pacemaker2: Fix distinfo
The distinfo now matches the previous one again. According to mat@
this toggling been noticed before but we never got to the bottom of it.
It might be a bug in Github's archive generation, I'll take a look
where and if I can report this.

This is a direct commit, as main has pacemaker 2.1.7

Fixes:	9bc5b3beb70b1ac01059b7952e692b6d7a6ff4ec