www/caddy{,-custom}: chown caddy.log to caddy user/group
We have set caddy.log to root:wheel, which works for the default case
of running caddy as root.

When caddy is run as a non-root users, caddy.log can't actually be
written to. Instead, chown caddy.log to the caddy user/group so that
it can be written to.

PR:		276150