multimedia/ptx-kmod: Deprecate and set expiration date to 2024-02-29
Legacy PCI-based hardware released around 2009 and 2010 that appears
to be obsolete. I tried to do research on this hardware and only
found a handful of pages mentioning it dating 10 years and older.
The hardware don't support the newer ISDB-S3 standard so I think
it's safe to assume that users have moved on.

PR:		276184
Approved by:	portmgr (maintainer timeout, 2+ weeks)