devel/jujutsu: fix install on other targets than amd64
This software uses a virtual manifest in its Cargo.toml, and it seems
that there is no generic way to deal with such Cargo.toml files,
unless building with direct access to the GitHub repository during the
install phase.

I'd appreciate any hint regarding a better way to install this program
to the stage directory than the hacked do-install target I use.