KDE: Update KDE Frameworks 5 to 5.115
Saturday, 10 February 2024

KDE today announces the release of KDE Frameworks 5.115.0.

KDE Frameworks are 83 addon libraries to Qt which provide a wide variety
of commonly needed functionality in mature, peer reviewed and well
tested libraries with friendly licensing terms. For an introduction see
the KDE Frameworks release announcement.

This release is part of a series of planned monthly releases making
improvements available to developers in a quick and predictable manner.
New in this version

PR:		277016
Exp-run by:	antoine