devel/llvm-devel: new snapshot
Bump major version to 19.

Add a build conflict with ${ARCH}-binutils for COMPILER_RT.  Technically
this is only an issue for 64-bit architectures that build 32-bit
sanitizers (the presence of <triple>-ld is picked up and it doesn't
pass some cmake test), but it's easier to just block all of them.