editors/openoffice-*: Fix build
Tweak our declaration for __cxa_allocate_exception() so that it matches
the one that sometimes shows up in /usr/include/c++/v1/cxxabi.h.  Clang
now treats the subtle differences as an error.

No PORTREVISION bump for this build fix.  There should not be any
functional change to the generated code if it was buildable before
this fix.