editors/openoffice-devel: Upgrade to a newer snapshot
Upgrade openoffice-devel to a newer snapshot with bug and
security fixes.

Remove the point release and snapshot timestamp from the suffix
appended to the names of the wrapper script and the program directory.
It is annoying when the suffix changes and we'll never support
the simultaneous installation of multiple versions where these
values differ.

patch-boost_boost__1__55__0.patch was upstreamed.

MFH:		2024Q1
Security:	CVE-2012-5639
Security:	CVE-2023-1183
Security:	CVE-2023-47804