devel/libiec61850: New port: open-source library for the IEC 61850 protocols
libiec61850 is an open-source (GPLv3) implementation of an IEC 61850 client and
server library implementing the protocols MMS, GOOSE and SV. It is implemented
in C (according to the C99 standard) to provide maximum portability. It can be
used to implement IEC 61850 compliant client and server applications on
embedded systems and PCs running FreeBSD, Linux, Windows, and MacOS. Included
is a set of simple example applications that can be used as a starting point to
implement own IEC 61850 compliant devices or to communicate with IEC 61850
devices. The library has been successfully used in many commercial software
products and devices.