math/py-pyprobables: Add py-pyprobables 0.6.0
pyprobables is a pure-python library for probabilistic data structures. The goal
is to provide the developer with a pure-python implementation of common
probabilistic data-structures to use in their work.

To achieve better raw performance, it is recommended supplying an alternative
hashing algorithm that has been compiled in C. This could include using the MD5
and SHA512 algorithms provided or installing a third party package and writing
your own hashing strategy. Some options include the murmur hash mmh3 or those
from the pyhash library. Each data object in pyprobables makes it easy to pass
in a custom hashing function.