textproc/py-spacy: Add py-spacy 3.7.4
spaCy is a library for advanced Natural Language Processing in Python and
Cython. It's built on the very latest research, and was designed from day one to
be used in real products.

spaCy comes with pretrained pipelines and currently supports tokenization and
training for 70+ languages. It features state-of-the-art speed and neural
network models for tagging, parsing, named entity recognition, text
classification and more, multi-task learning with pretrained transformers like
BERT, as well as a production-ready training system and easy model packaging,
deployment and workflow management.