textproc/py-spacy-legacy: Add py-spacy-legacy 3.0.12
spacy-legacy includes outdated registered functions for spaCy v3.x, for example
model architectures, pipeline components and utilities. It's installed
automatically as a dependency of spaCy, and allows us to provide backwards
compatibility, while keeping the core library tidy and up to date. All of this
happens under the hood, so you typically shouldn't have to care about this
package.