devel/llvm18: Update to 18.1.0rc3
Add a build conflict with ${ARCH}-binutils for COMPILER_RT to avoid
plist issues.  Technically this is only an issue for 64-bit
architectures that build 32-bit sanitizers (the presence of <triple>-ld
is picked up and it doesn't pass some cmake test), but it's easier to
just block all of them.