databases/pgpool-II-45: New port
pgpool is a connection pool server for PostgreSQL. pgpool runs between
PostgreSQL's clients(front ends) and servers(back ends). A PostgreSQL
client can connect to pgpool as if it were a standard PostgreSQL server.

pgpool caches the connection to PostgreSQL server to reduce the overhead
to establish the connection to it.

This is the 4.5.X series of the port.

WWW: https://pgpool.net