editors/imhex-current: Add new port
ImHex is a hex editor for reverse engineers and programmers.

This port is provided separately from existing editors/imhex port
for those who would like to use the latest version of ImHex.

The version requires FreeBSD 15-current, 14-stable, 13-stable,
or upcoming 13.3-RELEASE that has libc++ of LLVM 17 on the base
system to build it, thus it is not possible to build the version
on FreeBSD 13.2-RELEASE or 14.0-RELEASE. This situation has
prevented the update of editors/imhex port to the latest version.

This port is a tentative solution until FreeBSD 14.1-RELEASE
will be released.