emulators/wine8: New port, cloning emulators/wine
Wine 9 has been released some weeks ago and it's time to upgrade the
main emulators/wine port. However, there may be some regressions (in
particular around WoW64), so create emulators/wine8 as a clone of the
state before the version upgrade as a fallback. [1]

Since it is rarely, if ever, used, in particular in this context,
remove the MONO option.

Discussed with:	Lexi Winter <lexi@le-fay.org> [1]