Mk/Scripts/do-depends.sh: Avoid make -D flag
Both FreeBSD make and GNU make pass command line flags to sub-makes
through the MAKEFLAGS environment variable, but the flags aren't
compatible and GNU make 4.4 exits with an error when it encounters
flags it doesn't understand, e.g. -D.

Avoid using the flag for now until ports run gmake with SETENVI.

This fixes the case where a port is built using make directly and it
recurses to build a dependency that uses gmake.  Poudriere does not
use this feature of the ports tree to build dependencies so it's not
affected.

PR:		272216, 277492