emulators/qemu: Fix regression in gmake
This issue started popping up after the latest gmake upgrade. Although
there are multiple ways of fixing this I will stick with the preferred
path by tijl@ as he is trying to address the issue with other ports.

As I am not sure sure whether if this triggers a different binary being
generated but for the safest route bumping portrevision. This problem
is not reproducible for people using poudreiere to build their ports

PR:		277527
Reported by:	rozhuk.im@gmail.com