www/py-dj42-django-otp: Add py-dj42-django-otp 1.3.0 (copied from py-django-otp)
This project makes it easy to add support for one-time passwords (OTPs) to
Django. It can be integrated at various levels, depending on how much
customization is required. It integrates with django.contrib.auth, although it
is not a Django authentication backend. The primary target is developers wishing
to incorporate OTPs into their Django projects as a form of two-factor
authentication.

Several simple OTP plugins are included and more are available separately. This
package also includes an implementation of OATH HOTP and TOTP for convenience,
as these are standard OTP algorithms used by multiple plugins.