net/openldap26-server: fix build with lld 17 and DEBUG option off
If the DEBUG option for the net/openldap26-server is turned off,
building the port will result in:

  ld: error: version script assignment of 'OPENLDAP_2.200' to symbol 'ldap_dump_connection' failed: symbol not defined
  ld: error: version script assignment of 'OPENLDAP_2.200' to symbol 'ldap_dump_requests_and_responses' failed: symbol not defined
  cc: error: linker command failed with exit code 1 (use -v to see invocation)

This is because the mentioned symbols are in a #ifdef LDAP_DEBUG block,
and will indeed not be defined if DEBUG option is off.

Adjust the port Makefile to delete the undefined symbols from the
ldap.map file, if the DEBUG option is off.

PR:		277585
Reported by:	Matt Magoffin <bugs.freebsd.org@msqr.us>
MFH:		2024Q1