cad/openvsp: update to 3.37.2
https://github.com/OpenVSP/OpenVSP/releases/tag/OpenVSP_3.37.2

3.37.1:

This version changes a compiler flag that makes a new dependency build properly
on Windows. This in turn fixes CFDMesh and FEAMesh on Windows.  Without this
change, 3.37.0 will quickly crash if you try to use either mesher. This won’t
change anything for non-windows users.

3.37.2 contains bugfixes

For now, use bundled fltk since the port uses a newer version than the one in
ports.