net/realrek-re-kmod198: add port for 198 version
as a workaround for bug 275882. This port can be retired when the bug is
resolved completely.

Many people need the 198 version because of the hang-up issue. Another
set of people need 199 because of another issue. This port is needed to
satisfy both sets of people until complete until a complete solution for
275882 is found.

PR:		275882
Sponsored by:	Cybertrust Japan

(cherry picked from commit b770c919121526ebbf61b81fd6b832619319df60)