graphics/nvidia-drm-kmod: fix build depends and linuxkpi registration
This fixes an issue where linuxkpi/DRM compatibility was not being
correctly detected, causing sway to not work. The conftest.sh script
checks the installed kernel modules for symbol presence, but on
package builders drm.ko was not being installed due to being in
RUN_DEPENDS instead of BUILD_DEPENDS.

This also fixes a panic in sway when an external monitor is plugged
into a prime laptop. This uses linux_pci_attach_device when possible
to actually register everything needed for the pci dev, instead of
just filling in the bare miniumum.

Differential Revision: https://reviews.freebsd.org/D44308