net/py-haproxy-cli: new port
haproxy-cli is a tool to manage the various aspects of HAProxy that can be
controlled by means of its socket.

haproxy-cli can be used in 2 modes: CLI mode and Python API mode. CLI mode,
as the name implies, gives you a command, haproxy-cli, that can be used to
control HAProxy. You can use the Python API mode to integrate haproxy-cli
directly in your Python project.

PR:	275562