graphics/lux: fix noto dependency
Previous commit was not approved by maintainer.
Depending on the file ${GUIFONT} was intentional
as this file is hard-coded in the application.

Instead of removing this sanity check, the correct
is to adjust the dependency from the now meta port
to the port actually delivering this file.

Fixes:		 2370847ccdf81a5d6305bf6ec4e8f33765e6f75d.
(cherry picked from commit ee8fe6f4184ab83854c0581b43511d214bfc2730)